scan4halal-backend/requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
fsspec==2025.7.0
git-filter-repo==2.47.0
huggingface-hub==0.33.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.5
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.5.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==2.1.4
